+++ packages/ispect/lib/ispect.dart
@@ -16,6 +16,7 @@ export 'src/common/extensions/ispectify_data.dart';
 export 'src/common/models/panel_button.dart';
 export 'src/common/models/panel_item.dart';
 export 'src/common/observers/route_observer.dart';
+export 'src/common/services/logs_json_service.dart';
 export 'src/common/utils/file_utils.dart';
 export 'src/common/utils/logs_file/logs_file.dart';
 export 'src/common/widgets/builder/data_builder.dart';