+++ packages/ispect/lib/src/common/extensions/init.dart
@@ -31,6 +31,7 @@ extension ISpectifyFlutter on ISpectify {
     ISpectifyObserver? observer,
     ISpectifyOptions? options,
     ISpectifyFilter? filter,
+    ILogHistory? history,
   }) =>
       ISpectify(
         logger: (logger ?? ISpectifyLogger()).copyWith(
@@ -39,6 +40,7 @@ extension ISpectifyFlutter on ISpectify {
         options: options,
         observer: observer,
         filter: filter,
+        history: history,
       );
 
   /// Default output handler for logging in Flutter.
@@ -61,9 +63,10 @@ extension ISpectifyFlutter on ISpectify {
     switch (defaultTargetPlatform) {
       case TargetPlatform.iOS:
       case TargetPlatform.macOS:
-        log(message, name: 'ISpectify', level: 800); // INFO level logging
+        log(message, name: 'ISpect', level: 1000);
       default:
-        debugPrint(message);
+        // ignore: avoid_print
+        print(message);
     }
   }
 }