+++ packages/ispectify/lib/src/utils/console_utils.dart
@@ -38,34 +38,6 @@ abstract class ConsoleUtils {
     return withCorner ? '└$line' : line;
   }
 
-  /// Generates a top border line for console messages.
-  ///
-  /// This method returns a string consisting of a repeated `lineSymbol`
-  /// for the specified `length`. Optionally, a top-left corner ('┌')
-  /// can be added if `withCorner` is `true`.
-  ///
-  /// ### Example:
-  /// ```dart
-  /// print(ConsoleUtils.topLine(5)); // Output: ─────
-  /// print(ConsoleUtils.topLine(5, withCorner: true)); // Output: ┌─────
-  /// ```
-  ///
-  /// - `length`: The number of `lineSymbol` characters in the line.
-  /// - `lineSymbol`: The character used to construct the line. Defaults to '─'.
-  /// - `withCorner`: If `true`, adds a top-left corner symbol. Defaults to `false`.
-  ///
-  /// Returns:
-  /// A formatted string representing the top border line.
-  static String topLine(
-    int length, {
-    String lineSymbol = '─',
-    bool withCorner = false,
-  }) {
-    assert(length > 0, 'Line length must be greater than zero.');
-    final line = lineSymbol * length;
-    return withCorner ? '┌$line' : line;
-  }
-
   /// ANSI color mapping for different log levels.
   ///
   /// This map associates each `LogLevel` with a corresponding ANSI color.