+++ packages/ispectify_dio/lib/src/models/error.dart
@@ -61,6 +61,6 @@ class DioErrorLog extends ISpectifyData {
       buffer.write('\nHeaders: $prettyHeaders');
     }
 
-    return buffer.toString().truncated!;
+    return buffer.toString().truncate()!;
   }
 }